requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.5.0
Pillow>=9.0.0
opencv-python>=4.7.0
numpy>=1.24.0
pytesseract>=0.3.10
